backend/requirements.txt
fastapi
uvicorn
requests
beautifulsoup4
perplexityai
python-multipart
pillow
python-dotenv
openai
numpy
google-generativeai
pika
redis
httpx
